+++ commercetools/commercetools-sdk-java-api/src/main/java-generated/com/commercetools/api/models/shipping_method/ShippingMethodDraftImpl.java
@@ -37,6 +37,8 @@ public class ShippingMethodDraftImpl implements ShippingMethodDraft, ModelBase {
 
     private java.util.List<com.commercetools.api.models.shipping_method.ZoneRateDraft> zoneRates;
 
+    private Boolean active;
+
     private Boolean isDefault;
 
     private String predicate;
@@ -53,7 +55,8 @@ public class ShippingMethodDraftImpl implements ShippingMethodDraft, ModelBase {
             @JsonProperty("localizedDescription") final com.commercetools.api.models.common.LocalizedString localizedDescription,
             @JsonProperty("taxCategory") final com.commercetools.api.models.tax_category.TaxCategoryResourceIdentifier taxCategory,
             @JsonProperty("zoneRates") final java.util.List<com.commercetools.api.models.shipping_method.ZoneRateDraft> zoneRates,
-            @JsonProperty("isDefault") final Boolean isDefault, @JsonProperty("predicate") final String predicate,
+            @JsonProperty("active") final Boolean active, @JsonProperty("isDefault") final Boolean isDefault,
+            @JsonProperty("predicate") final String predicate,
             @JsonProperty("custom") final com.commercetools.api.models.type.CustomFieldsDraft custom) {
         this.key = key;
         this.name = name;
@@ -62,6 +65,7 @@ public class ShippingMethodDraftImpl implements ShippingMethodDraft, ModelBase {
         this.localizedDescription = localizedDescription;
         this.taxCategory = taxCategory;
         this.zoneRates = zoneRates;
+        this.active = active;
         this.isDefault = isDefault;
         this.predicate = predicate;
         this.custom = custom;
@@ -130,7 +134,15 @@ public class ShippingMethodDraftImpl implements ShippingMethodDraft, ModelBase {
     }
 
     /**
-     *  <p>If <code>true</code> the ShippingMethod will be the Project's default ShippingMethod.</p>
+     *  <p>If set to <code>true</code>, the ShippingMethod can be used during the creation or update of a Cart or Order.</p>
+     */
+
+    public Boolean getActive() {
+        return this.active;
+    }
+
+    /**
+     *  <p>If set to <code>true</code>, the ShippingMethod will be the Project's default ShippingMethod.</p>
      */
 
     public Boolean getIsDefault() {
@@ -189,6 +201,10 @@ public class ShippingMethodDraftImpl implements ShippingMethodDraft, ModelBase {
         this.zoneRates = zoneRates;
     }
 
+    public void setActive(final Boolean active) {
+        this.active = active;
+    }
+
     public void setIsDefault(final Boolean isDefault) {
         this.isDefault = isDefault;
     }
@@ -218,6 +234,7 @@ public class ShippingMethodDraftImpl implements ShippingMethodDraft, ModelBase {
                 .append(localizedDescription, that.localizedDescription)
                 .append(taxCategory, that.taxCategory)
                 .append(zoneRates, that.zoneRates)
+                .append(active, that.active)
                 .append(isDefault, that.isDefault)
                 .append(predicate, that.predicate)
                 .append(custom, that.custom)
@@ -228,6 +245,7 @@ public class ShippingMethodDraftImpl implements ShippingMethodDraft, ModelBase {
                 .append(localizedDescription, that.localizedDescription)
                 .append(taxCategory, that.taxCategory)
                 .append(zoneRates, that.zoneRates)
+                .append(active, that.active)
                 .append(isDefault, that.isDefault)
                 .append(predicate, that.predicate)
                 .append(custom, that.custom)
@@ -243,6 +261,7 @@ public class ShippingMethodDraftImpl implements ShippingMethodDraft, ModelBase {
                 .append(localizedDescription)
                 .append(taxCategory)
                 .append(zoneRates)
+                .append(active)
                 .append(isDefault)
                 .append(predicate)
                 .append(custom)
@@ -258,6 +277,7 @@ public class ShippingMethodDraftImpl implements ShippingMethodDraft, ModelBase {
                 .append("localizedDescription", localizedDescription)
                 .append("taxCategory", taxCategory)
                 .append("zoneRates", zoneRates)
+                .append("active", active)
                 .append("isDefault", isDefault)
                 .append("predicate", predicate)
                 .append("custom", custom)