+++ changes.md
@@ -100,6 +100,8 @@
 - added property `variants` to type `ProductTailoringData`
 - added property `variants` to type `ProductTailoringDraft`
 - added property `variants` to type `ProductTailoringInStoreDraft`
+- added property `active` to type `ShippingMethod`
+- added property `active` to type `ShippingMethodDraft`
 - added property `store` to type `StagedQuote`
 </details>
 
@@ -156,6 +158,7 @@
 - added type `ProductTailoringSetImageLabelAction`
 - added type `CustomerSearchStatus`
 - added type `ProjectChangeCustomerSearchStatusAction`
+- added type `ShippingMethodChangeActiveAction`
 </details>